Moved SearchResults to molecules folder
