+++ src/components/Molecules/SearchResults/SearchResults.interface.ts
