+++ src/components/Molecules/SearchResults/SearchResults.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/global/tokens" as *;
+@use "../../../styles/global/tokens" as *;
 
 .search-results-bg {
   background-color: $color-grey-100;
